+++ DBManager/Reapack_Header.lua
@@ -1,17 +1,27 @@
 --[[
 @metapackage true
 @description DBManager
-@version 0.8beta
+@version 0.81beta
 @author Claudiohbsantos
 @link http://claudiohbsantos.com
-@date 2019 08 06
+@date 2019 08 19
 @about
     Database manager GUI for Reaper Media Explorer.
     Uses DBAssistant to make changes - https://github.com/Claudiohbsantos/DBAssistant . The executable will be automatically downloaded upon installation. 
     Depends on JS Reascript API and Lokasenna GUI Library v2.
 @changelog
-    - Includes DBAssistant 0.3.4
-    - fixed parent detection
+    # DBManager v0.81beta
+
+    ## BUGS
+
+    - error when searching without making selection
+    - false error dialogue when importing config file
+    - mismatched references  on list after save
+    - fixed preview copy destination on import sfx window
+
+    ## IMPROVEMENTS
+
+    - Single popup for new DB creation
 @provides
     [main] DBManager.lua > DBManager.lua
     [nomain] lua_modules/json.lua > lua_modules/json.lua